Misc fixes: - Fixed memory corruption when a thread dies in a very unlikely likely way (after starting to exit) - Created new voltage monitor soruce module which reads from USB device - Fix bug in fifo buffer which corrupted it when inserting ordered - Write average calculation and buffer entry timeout for averager

+++ src/lib/measurement.h
@@ -28,7 +28,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "../lib/messages.h"
 
 struct vl_reading {
-	uint64_t reading_millis;
+	unsigned int reading_millis;
 	struct vl_message message;
 };
